feat: Add molecule class with methods to add atoms and liaisons
This commit adds a new `Molecule` class to the codebase. The `Molecule` class has two new methods: `add_atom` and `add_liaison`. These methods allow the user to add atoms and liaisons to the molecule.

The `add_atom` method takes a `type` parameter and creates a new `Atome` object with the specified type. The `Atome` object is then added to the `atomes` list of the `Molecule` object.

The `add_liaison` method takes two `atom` parameters and creates a new `Liaison` object between the two atoms. The `Liaison` object is then added to the `liaisons` list of the `Molecule` object.

This commit improves the codebase by providing a convenient way to add atoms and liaisons to a molecule.